Fix: lint and rename simple user to match back
