+++ lib/flappybird/providers/bird_provider.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:myecl/flappybird/class/bird.dart';
 import 'package:myecl/flappybird/providers/bird_image_provider.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 import 'package:myecl/user/providers/user_provider.dart';
 
 class BirdNotifier extends StateNotifier<Bird> {
