+++ lib/raffle/class/tickets.dart
@@ -1,6 +1,6 @@
 import 'package:myecl/raffle/class/pack_ticket.dart';
 import 'package:myecl/raffle/class/prize.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 
 class Ticket {
   Ticket({
