+++ lib/raffle/ui/pages/admin_module_page/adding_user_card.dart
@@ -4,7 +4,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:myecl/raffle/class/cash.dart';
 import 'package:myecl/raffle/providers/cash_provider.dart';
 import 'package:myecl/raffle/tools/constants.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 
 class AddingUserCard extends HookConsumerWidget {
   final SimpleUser user;
