+++ lib/user/class/user.dart
@@ -2,7 +2,7 @@ import 'package:myecl/admin/class/account_type.dart';
 import 'package:myecl/admin/class/simple_group.dart';
 import 'package:myecl/tools/functions.dart';
 import 'package:myecl/user/class/applicant.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 import 'package:myecl/user/class/floors.dart';
 
 class User {
