+++ lib/vote/class/members.dart
@@ -1,6 +1,6 @@
 import 'package:myecl/admin/class/account_type.dart';
 import 'package:myecl/tools/functions.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 
 class Member extends SimpleUser {
   late String role;
