+++ test/admin/group_provider_test.dart
@@ -4,7 +4,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:myecl/admin/class/group.dart';
 import 'package:myecl/admin/providers/group_provider.dart';
 import 'package:myecl/admin/repositories/group_repository.dart';
-import 'package:myecl/user/class/list_users.dart';
+import 'package:myecl/user/class/simple_users.dart';
 
 class MockGroupRepository extends Mock implements GroupRepository {}
